move tfunc to compilers
